// app/requests/_personal/postConvertTask/postConvertTaskSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PostConvertTaskSuccess {
  data: {
    id: number;
    selfEmployed: boolean;
    status: 1 | 2 | 3 | 4 | 5;
    place: {
      id: number;
      name: string;
      latitude: string;
      longitude: string;
      address_kladr: string;
      logo: string;
      region: {
        id: number;
        name: string;
        [k: string]: unknown;
      };
      brand: {
        id: number;
        name: string;
        logo: string;
        description: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    project: {
      id: number;
      name: string;
      brand: {
        id: number;
        name: string;
        logo: string;
        description: string;
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    user: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name: string;
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    acceptUser: null | {
      id: number;
      phone: number;
      email: string;
      logo: string;
      [k: string]: unknown;
    };
    orderActivities: {
      id: number;
      viewActivity: {
        id: number;
        name: string;
        detailName: string;
        previewText: string;
        logo: string;
        traveling: boolean;
        [k: string]: unknown;
      };
      count: number;
      dateStart: string;
      dateEnd: string;
      needFoto: boolean;
      dateActivity: {
        timeStart: string;
        timeEnd: string;
        places: {
          id: number;
          name: string;
          latitude: string;
          longitude: string;
          address_kladr: string;
          logo?: null | string;
          region: {
            id: number;
            name: string;
            [k: string]: unknown;
          };
          brand: {
            id: number;
            name: string;
            logo: null | string;
            description: string;
            [k: string]: unknown;
          };
          [k: string]: unknown;
        }[];
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    }[];
    acceptedUser: {
      id: number;
      phone: number;
      email: string;
      logo: string;
      roles: {
        id: number;
        name:
          | "admin"
          | "client"
          | "manager"
          | "recruiter"
          | "specialist"
          | "supervisor";
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}